fix: Update Gatsby dependencies and config
This upgrades our Gatsby and Gatsby plugin versions to their latest
versions, so that we're good on security updates and standard practices
across other repos.

Update configuration to match the new modules' configuration.

Since I had to touch the config for app manifest and the icons I updated
these to match the project details.