feat: Implement client server websocket comms
The client drawing app now opens a WebSocket to the server and sends
drawing data.

This required updating how Gatsby renders the drawing app, using
client-rendering only. We also changed the server to use Express so that
we can manage CORS between the two instances.

Right now this disabled the Arduino communication. Passing it off to
Kate to work on that next.